pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragondin"
version = "0.1.0"
description = "Local personal RAG CLI"
authors = [
   { name = "Augustin Bresset" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "click",
    "faiss-cpu",
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain_huggingface",
    "langchain-text-splitters",
    "sentence-transformers",
]

[project.scripts]
ragondin = "ragondin.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]